removed is xray otlp endpoint validation in the span exporter
